+++ pkg/textual/chain.go
@@ -37,11 +37,11 @@ import (
 //
 // See also: Router for fan-out/fan-in routing and SyncApply for one-shot
 // processing.
-type Chain[S UTF8Stringer[S]] struct {
+type Chain[S Carrier[S]] struct {
 	processors []Processor[S]
 }
 
-func NewChain[S UTF8Stringer[S]](processors ...Processor[S]) *Chain[S] {
+func NewChain[S Carrier[S]](processors ...Processor[S]) *Chain[S] {
 	return &Chain[S]{
 		processors: processors,
 	}