+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 + Using `textual.String` in the [reverse_words example](examples/reverse_words/README.md).
-+ Added `textual.String` a minimal generic implementation of `textual.UTF8Stringer`
-+ Adapted `textual.Result` to implement `textual.UTF8Stringer`
-+ Exposing a generic interface `textual.UTF8Stringer`, and refactored the stack.
++ Added `textual.String` a minimal generic implementation of `textual.Carrier`
++ Adapted `textual.Result` to implement `textual.Carrier`
++ Exposing a generic interface `textual.Carrier`, and refactored the stack.
 + Added `SyncApply`..
 + `textual.UTF8String` is now a symbolic string alias.
 + [Textual.swift](helpers/swift/README.md) a Lightweight Swift utility to work with `textual` `Result` objects in iOS / macOS clients.